--- packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/NewSingleImageViewer/index.js
@@ -1,1 +1,0 @@
-export { default } from './NewSingleImageViewerConnector'
