+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SingleImageViewer/SingleImageViewer.js
@@ -1,112 +1,157 @@
+import { ParentSize } from '@visx/responsive'
 import { Box } from 'grommet'
-import PropTypes from 'prop-types'
-import { useRef } from 'react'
-import styled, { css } from 'styled-components'
+import { bool, func, number, shape, string } from 'prop-types'
+import { useEffect, useRef } from 'react'
 
 import SVGContext from '@plugins/drawingTools/shared/SVGContext'
+
 import InteractionLayer from '../InteractionLayer'
 import ZoomControlButton from '../ZoomControlButton'
-import locationValidator from '../../helpers/locationValidator'
 
-const PlaceholderSVG = styled.svg`
-  background: no-repeat center / cover url('https://static.zooniverse.org/www.zooniverse.org/assets/fe-project-subject-placeholder-800x600.png');
-  touch-action: pinch-zoom;
-  max-width: ${props => props.$maxWidth || '100%'};
-  ${props => props.$maxHeight && css`max-height: ${props.$maxHeight};`}
-`
-const SVGImageCanvas = styled.svg`
-  overflow: visible;
-`
+import SVGImage from '../SVGComponents/SVGImage'
+import VisXZoom from '../SVGComponents/VisXZoom'
+
+const DEFAULT_HANDLER = () => true
+
 function SingleImageViewer({
-  children,
-  enableInteractionLayer = false,
+  enableInteractionLayer = true,
+  enableRotation = DEFAULT_HANDLER,
   frame = 0,
-  height,
+  imgRef,
+  invert = false,
   limitSubjectHeight = false,
-  onKeyDown = () => true,
-  rotate = 0,
-  svgMaxHeight = null,
+  move = false,
+  naturalHeight,
+  naturalWidth,
+  panning = true,
+  rotation = 0,
+  scale = 1,
+  setOnZoom = DEFAULT_HANDLER,
+  setOnPan = DEFAULT_HANDLER,
+  src,
+  subject,
+  subjectId,
   title = {},
-  viewBox,
-  width,
-  zoomControlFn = null,
-  zooming = false
+  zoomControlFn,
+  zooming = true
 }) {
   const canvasLayer = useRef()
   const canvas = canvasLayer.current
-  const transform = `rotate(${rotate} ${width / 2} ${height / 2})`
+
+  useEffect(function onMount() {
+    enableRotation()
+  }, [])
+
+  const rotationTransform = `rotate(${rotation} ${naturalWidth / 2} ${naturalHeight / 2})`
+
+  const viewBox = `0 0 ${naturalWidth} ${naturalHeight}`
+
+  const SVGImageComponent = ({
+    children,
+    transform: panZoomTransform
+  }) => (
+    <svg
+      ref={canvasLayer}
+      viewBox={viewBox}
+    >
+      <g transform={panZoomTransform}>
+        <SVGImage
+          ref={imgRef}
+          invert={invert}
+          move={move}
+          naturalHeight={naturalHeight}
+          naturalWidth={naturalWidth}
+          src={src}
+          subjectID={subjectId}
+        />
+        {children}
+        {enableInteractionLayer && (
+          <InteractionLayer
+            frame={frame}
+            height={naturalHeight}
+            scale={scale}
+            subject={subject}
+            width={naturalWidth}
+          />
+        )}
+      </g>
+    </svg>
+  )
 
   return (
-    <SVGContext.Provider value={{ canvas, viewBox, rotate, width, height }}>
-      {zoomControlFn && (
-        <ZoomControlButton
-          onClick={zoomControlFn}
-          zooming={zooming}
-        />
-      )}
-      <Box
-        animation='fadeIn'
-        overflow='hidden'
-        width='100%'
-        align='flex-end'
-      >
-        <PlaceholderSVG
-          focusable
-          $maxHeight={svgMaxHeight}
-          $maxWidth={limitSubjectHeight ? `${width}px` : '100%'}
-          onKeyDown={onKeyDown}
-          tabIndex={0}
-          viewBox={`0 0 ${width} ${height}`}
+    <ParentSize>
+      {(parent) => (
+        <SVGContext.Provider
+          value={{
+            canvas,
+            viewBox,
+            rotate: rotation,
+            width: parent.width,
+            height: parent.height
+          }}
         >
-          {title?.id && title?.text && (
-            <title id={title.id}>{title.text}</title>
-          )}
-          <g
-            transform={transform}
+          <Box
+            align='flex-end'
+            animation='fadeIn'
+            height={parent.height}
+            overflow='hidden'
+            width={parent.width}
           >
-            <SVGImageCanvas
-              ref={canvasLayer}
-              viewBox={viewBox}
-            >
-              {children}
-              {enableInteractionLayer && (
-                <InteractionLayer
-                  frame={frame}
-                  height={height}
-                  width={width}
+            {zoomControlFn &&
+              <ZoomControlButton
+                onClick={zoomControlFn}
+                position='absolute'
+                zooming={zooming}
+              />}
+            <svg viewBox={viewBox}>
+              <g transform={rotationTransform}>
+                <VisXZoom
+                  height={naturalHeight}
+                  panning={panning}
+                  setOnPan={setOnPan}
+                  setOnZoom={setOnZoom}
+                  width={naturalWidth}
+                  zoomingComponent={SVGImageComponent}
+                  zooming={zooming}
                 />
-              )}
-            </SVGImageCanvas>
-          </g>
-        </PlaceholderSVG>
-      </Box>
-    </SVGContext.Provider>
+              </g>
+            </svg>
+          </Box>
+        </SVGContext.Provider>
+      )}
+    </ParentSize>
   )
 }
 
 SingleImageViewer.propTypes = {
-  /** Index of the Frame. Initially inherits from parent Viewer or overwritten in Viewer with SubjectViewerStore */
-  frame: PropTypes.number,
-  /** Passed from container */
-  enableInteractionLayer: PropTypes.bool,
-  /** Calculated by useSubjectImage() */
-  height: PropTypes.number.isRequired,
-  /** Stored in subject viewer store */
-  rotate: PropTypes.number,
-  /** Calculated in SVGPanZoom component */
-  svgMaxHeight: PropTypes.string,
-  title: PropTypes.shape({
-    id: PropTypes.string,
-    text: PropTypes.string
+  enableInteractionLayer: bool,
+  enableRotation: func,
+  frame: number,
+  imgRef: shape({
+    current: shape({
+      naturalHeight: number,
+      naturalWidth: number,
+      src: string
+    })
   }),
-  /** Calculated in SVGPanZoom component */
-  viewBox: PropTypes.string,
-  /** Calculated by useSubjectImage() */
-  width: PropTypes.number.isRequired,
-  /** Stored in subject viewer store */
-  zoomControlFn: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-  /** Passed from container */
-  zooming: PropTypes.bool
+  invert: bool,
+  limitSubjectHeight: bool,
+  move: bool,
+  naturalHeight: number,
+  naturalWidth: number,
+  panning: bool,
+  rotation: number,
+  scale: number,
+  setOnZoom: func,
+  setOnPan: func,
+  src: string,
+  subjectId: string,
+  title: shape({
+    id: string,
+    text: string
+  }),
+  zoomControlFn: func,
+  zooming: bool
 }
 
 export default SingleImageViewer
