+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SingleImageViewer/SingleImageViewer.stories.js
@@ -1,51 +1,68 @@
 import { Box } from 'grommet'
 import { Provider } from 'mobx-react'
-import SubjectViewerStore from '@store/SubjectViewerStore'
-import SingleImageViewer from '@viewers/components/SingleImageViewer'
-import asyncStates from '@zooniverse/async-states'
+
+import ImageToolbar from '../../../ImageToolbar'
 
 import mockStore from '@test/mockStore'
 import { SubjectFactory, WorkflowFactory } from '@test/factories'
+
+import SingleImageViewerConnector from '.'
 
 const subject = SubjectFactory.build({
   locations: [{ 'image/jpeg': 'https://panoptes-uploads.zooniverse.org/production/subject_location/11f98201-1c3f-44d5-965b-e00373daeb18.jpeg' }]
 })
 
-const workflow = WorkflowFactory.build('workflow', {
+const workflow = WorkflowFactory.build({
   configuration: {
-    limit_subject_height: true
+    invert_subject: true
   }
 })
 
+const store = mockStore({ subject, workflow })
+
+const ViewerContext = ({ store, children }) => {
+  return <Provider classifierStore={store}>{children}</Provider>
+}
+
 export default {
   title: 'Subject Viewers / SingleImageViewer',
-  component: SingleImageViewer
+  component: SingleImageViewerConnector
 }
 
 export function Default() {
   return (
-    <Provider classifierStore={Default.store}>
+    <ViewerContext store={store}>
       <Box width='large'>
-        <SingleImageViewer
-          loadingState={asyncStates.success}
-          enableInteractionLayer={false}
-        />
+        <SingleImageViewerConnector />
       </Box>
-    </Provider>
+    </ViewerContext>
   )
 }
-Default.store = mockStore({ subject })
+
+export function PanAndZoom() {
+  return (
+    <ViewerContext store={store}>
+      <Box direction='row' width='large'>
+        <SingleImageViewerConnector />
+        <ImageToolbar width='4rem' />
+      </Box>
+    </ViewerContext>
+  )
+}
+
+const workflowWithLimitSubjectHeight = WorkflowFactory.build({
+  configuration: {
+    limit_subject_height: true
+  }
+})
+const storeWithLimitSubjectHeight = mockStore({ subject, workflow: workflowWithLimitSubjectHeight })
 
 export function LimitSubjectHeight() {
   return (
-    <Provider classifierStore={LimitSubjectHeight.store}>
+    <ViewerContext store={storeWithLimitSubjectHeight}>
       <Box width='large'>
-        <SingleImageViewer
-          loadingState={asyncStates.success}
-          enableInteractionLayer={false}
-        />
+        <SingleImageViewerConnector />
       </Box>
-    </Provider>
+    </ViewerContext>
   )
 }
-LimitSubjectHeight.store = mockStore({ subject, workflow })
