+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/SingleImageViewer/index.js
@@ -1,3 +1,1 @@
 export { default } from './SingleImageViewerConnector'
-export { default as SingleImageViewerContainer } from './SingleImageViewerContainer'
-export { default as SingleImageViewer } from './SingleImageViewer'
